+++ packages/flutter_robusta/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.2.0
-- BC: replace `MaterialExtension` and `CupertinoExtension` with `AppExtension`.
+- BC: replace `MaterialExtension` and `CupertinoExtension` with `FlutterAppExtension`.
 - Support settings `AppExtension` via `Configurator` extension.
 
 ## 0.1.0+1